ARGH, uint256 for checkpoint.

+++ src/checkpoints.cpp
@@ -27,7 +27,7 @@ namespace Checkpoints
     static MapCheckpoints mapCheckpoints =
         boost::assign::map_list_of
         ( 0, hashGenesisBlockOfficial )
-        ( 100, uint64("0x0000035c3bd074ada5f0a70d91c9e1d4a03bcac7c59470fa29f6b3f2fb084189"))
+        ( 100, uint256("0x0000035c3bd074ada5f0a70d91c9e1d4a03bcac7c59470fa29f6b3f2fb084189"))
         ;
 
     bool CheckHardened(int nHeight, const uint256& hash)
